Alteracao da representacao binaria para array de inteiros

+++ src/main/java/br/usp/ia/logic/mutation/impl/SimpleMutation.java
@@ -21,7 +21,7 @@ public class SimpleMutation implements Mutation {
         if (this.random.getUniformGenerator().nextDouble() <= probability) {
 
             //Calcula aleatoriamente o novo valor para o gene
-            final byte gene = (byte) this.random.getUniformGenerator().nextInt(2);
+            final int gene = this.random.getUniformGenerator().nextInt(2);
             individual.setGene(changingGene, gene);
         }
